+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of nrpe
 
+## Unreleased
+
+- Remove windows references and use Chef defaults for `file` and `template` resources
+
 ## 2.0.5 (2018-12-14)
 
 - Backport systemd unit from the package install to the cookbook